pseudocodigo hasta punto 3 del proyecto

+++ Pruebas/test5.pas
@@ -1,4 +1,4 @@
-FUN main( arg:INT, arg2:FLOAT )
+FUN main( arg:INT, arg2:INT )
 a:INT[4];
 BEGIN
 	if arg > 0 and arg2 < 3 then
